requirements-devel.txt
-r requirements.txt

factory_boy==2.5.1
py==1.4.26
pytest==2.7.0
pytest-django==2.8.0
pytest-pythonpath==0.6

coverage==3.7.1
coveralls==0.4.2
django-slowdown==0.0.1

transifex-client==0.11.1.beta
